update: improve source registry retrieval
